requirements.txt
Flask>=3.0.3
Werkzeug>=2.2.0
Flask-SQLAlchemy>=3.1.1
mysql-connector-python>=8.0.23